+++ src/pdl_api.h
@@ -12,5 +12,5 @@ void PDLBlit(uint32_t *bitmap);
 uint8_t PDLGetKey(uint8_t key);
 
 bool PDLDoWindowMessages();
-bool PDLInit(const char *title, size_t title_len, uint32_t x, uint32_t y,
+bool PDLInit(const char *title, uint32_t title_len, uint32_t x, uint32_t y,
              uint32_t width, uint32_t height);
